+++ .github/CODEOWNERS
@@ -154,3 +154,6 @@ Lib/ast.py                    @isidentical
 
 # pathlib
 **/*pathlib*                  @brettcannon
+
+# zipfile.Path
+**/*zipfile/*_path.py         @jaraco
