+++ Misc/NEWS.d/next/Library/2022-10-08-15-41-00.gh-issue-98098.DugpWi.rst
@@ -1,0 +1,2 @@
+Created packages from zipfile and test_zipfile modules, separating
+``zipfile.Path`` functionality.
